Put `O1TimedReply` into separate source files.
